vcmi: split bonus updaters
