+++ lib/IGameCallback.cpp
@@ -19,6 +19,7 @@
 #include "BattleFieldHandler.h"
 #include "ObstacleHandler.h"
 #include "bonuses/ILimiter.h"
+#include "bonuses/IUpdater.h"
 
 #include "serializer/CSerializer.h" // for SAVEGAME_MAGIC
 #include "serializer/BinaryDeserializer.h"
