+++ lib/battle/BattleInfo.cpp
@@ -9,7 +9,8 @@
  */
 #include "StdInc.h"
 #include "BattleInfo.h"
-#include "../bonuses/ILimiter.h"
+#include "bonuses/ILimiter.h"
+#include "bonuses/IUpdater.h"
 #include "../CStack.h"
 #include "../CHeroHandler.h"
 #include "../NetPacks.h"
